+++ views/default/components/list/require.php
@@ -1,5 +1,2 @@
-<script>
-	require(['jquery', 'components/list/init'], function($) {
-		$('.elgg-list,.elgg-gallery,.elgg-no-results').trigger('initialize');
-	});
-</script>
+<?php
+elgg_import_esm('js/components/list/init');